backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
Django==4.1.6
django-cors-headers==3.13.0
django-phonenumber-field==7.0.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
idna==3.4
phonenumbers==8.13.6
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
twilio==7.16.3
tzdata==2022.7
urllib3==1.26.14
